requirements.txt
ollama
python-docx
unstructured[docx]
